# docs/swagger.yaml
basePath: /
definitions:
  dto.ContactRequest:
    properties:
      email:
        type: string
      message:
        type: string
      name:
        type: string
      phone:
        type: string
    required:
    - email
    - message
    - name
    - phone
    type: object
  dto.CreateAboutSectionRequest:
    properties:
      content: {}
      description:
        type: string
      image:
        type: string
      is_active:
        type: boolean
      section_key:
        type: string
      sort_order:
        type: integer
      subtitle:
        type: string
      title:
        type: string
    required:
    - section_key
    type: object
  dto.CreateFAQRequest:
    properties:
      answer:
        type: string
      category:
        type: string
      is_active:
        type: boolean
      question:
        type: string
      sort_order:
        type: integer
    required:
    - answer
    - question
    type: object
...
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/dto.UserInfo'
    type: object
  dto.UpdateAboutSectionRequest:
    properties:
      content: {}
      description:
        type: string
      image:
        type: string
      is_active:
        type: boolean
      section_key:
        type: string
      sort_order:
        type: integer
      subtitle:
        type: string
      title:
        type: string
    type: object
  dto.UpdateFAQRequest:
    properties:
      answer:
        type: string
      category:
        type: string
      is_active:
        type: boolean
      question:
        type: string
      sort_order:
        type: integer
    type: object
...
  dto.UpdateInquiryStatusRequest:
    properties:
      status:
        type: string
    required:
    - status
    type: object
...
  dto.UserInfo:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        type: string
    type: object
  models.AboutSection:
    properties:
      content: {}
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image:
        type: string
      is_active:
        type: boolean
      section_key:
        type: string
      sort_order:
        type: integer
      subtitle:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.ContactInquiry:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      message:
        type: string
      name:
        type: string
      phone:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  models.FAQ:
    properties:
      answer:
        type: string
      category:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      question:
        type: string
      sort_order:
        type: integer
      updated_at:
        type: string
    type: object
...
  services.DashboardStats:
    properties:
      total_contact_inquiries:
        type: integer
      total_loan_inquiries:
        type: integer
      total_services:
        type: integer
      total_testimonials:
        type: integer
    type: object
  utils.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  utils.PaginatedResponse:
    properties:
      data: {}
      page:
        type: integer
      page_size:
        type: integer
      success:
        type: boolean
      total:
        type: integer
      total_pages:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API server for GK Capital finance website.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: GK Capital API Documentation
  version: "1.0"
paths:
...
  /api/admin/about:
    get:
      description: Get paginated list of all about sections
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.PaginatedResponse'
      security:
      - BearerAuth: []
      summary: Get all about sections for administration
      tags:
      - Admin About Sections
    post:
      consumes:
      - application/json
      description: Create an about section
      parameters:
      - description: Create About Section Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAboutSectionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.AboutSection'
              type: object
      security:
      - BearerAuth: []
      summary: Create a new about section
      tags:
      - Admin About Sections
  /api/admin/about/{id}:
    delete:
      description: Delete an about section by ID
      parameters:
      - description: About Section ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete an about section
      tags:
      - Admin About Sections
    put:
      consumes:
      - application/json
      description: Update details of an about section by ID
      parameters:
      - description: About Section ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update About Section Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateAboutSectionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.AboutSection'
              type: object
      security:
      - BearerAuth: []
      summary: Update an existing about section
      tags:
      - Admin About Sections
  /api/admin/contact-inquiries:
    get:
      description: Get list of contact inquiries with search and status filters
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Filter by status (new/read/responded)
        in: query
        name: status
        type: string
      - description: Search query (name, email, phone)
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.PaginatedResponse'
      security:
      - BearerAuth: []
      summary: Get contact inquiries
      tags:
      - Admin Inquiries
  /api/admin/contact-inquiries/{id}:
    patch:
      consumes:
      - application/json
      description: Update status of an inquiry by ID
      parameters:
      - description: Inquiry ID
        in: path
        name: id
        required: true
        type: integer
      - description: New status request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateInquiryStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Update contact inquiry status
      tags:
      - Admin Inquiries
  /api/admin/dashboard:
    get:
      description: Fetch aggregated metrics for services, testimonials, and inquiries
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/services.DashboardStats'
              type: object
      security:
      - BearerAuth: []
      summary: Get Admin Dashboard Statistics
      tags:
      - Admin Dashboard
  /api/admin/faqs:
    get:
      description: Get paginated list of all FAQs including inactive ones
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.PaginatedResponse'
      security:
      - BearerAuth: []
      summary: Get all FAQs for administration
      tags:
      - Admin FAQs
    post:
      consumes:
      - application/json
      description: Create a FAQ for administration
      parameters:
      - description: Create FAQ Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateFAQRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.FAQ'
              type: object
      security:
      - BearerAuth: []
      summary: Create a new FAQ
      tags:
      - Admin FAQs
  /api/admin/faqs/{id}:
    delete:
      description: Delete a FAQ by ID
      parameters:
      - description: FAQ ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete a FAQ
      tags:
      - Admin FAQs
    put:
      consumes:
      - application/json
      description: Update details of a FAQ by ID
      parameters:
      - description: FAQ ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update FAQ Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateFAQRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.FAQ'
              type: object
      security:
      - BearerAuth: []
      summary: Update an existing FAQ
      tags:
      - Admin FAQs
...
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate admin user and return JWT token
      parameters:
      - description: Login Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Admin Login
      tags:
      - Auth
  /api/contact:
    post:
      consumes:
      - application/json
      description: Submit the public contact form
      parameters:
      - description: Contact inquiry submission data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ContactRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ContactInquiry'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Submit a contact inquiry
      tags:
      - Contact Inquiry
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"